Add cookie support to ovs_lib, ovs_neutron_agent
All of the flows on the integration bridge belonging to a vif's
OpenFlow port are removed on port_bound. Since the firewall is invoked
to setup port filters before connectivity is established (i.e. before
port_bound is called) this defeats an OVS-based firewall. To solve
this issue, a distinct cookie on flows that belong to the firewall
and a separate, distinct cookie on flows that belong to the agent
will be used. This will require OVS version 1.5.0+ to delete flows
based on a cookie.

DocImpact: Adds configuration parameter to ovs_neutron_plugin.ini

Change-Id: Ia9b3aa6a383c3465f595725442eca2db05f24a16
Partially-implements: blueprint ovs-firewall-driver